requirements.txt
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
psycopg2-binary==2.7.4
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.4
whitenoise==3.3.1
